+++ packages/cadl-ranch-expect/src/decorators.ts
@@ -11,7 +11,7 @@ import {
   Program,
   StringLiteral,
 } from "@typespec/compiler";
-import { $route, $server, getOperationVerb, getRoutePath, HttpVerb } from "@typespec/http";
+import { $route, $server, getOperationVerb, getRoutePath, getServers, HttpVerb } from "@typespec/http";
 import { $versioned } from "@typespec/versioning";
 import { reportDiagnostic } from "./lib.js";
 import { SupportedBy } from "./types.js";
@@ -86,11 +86,7 @@ export interface ScenarioEndpoint {
 }
 
 export function listScenarios(program: Program): Scenario[] {
-  const serviceNamespace = listServices(program)[0].type;
-  if (serviceNamespace === undefined) {
-    return [];
-  }
-  return listScenarioIn(program, serviceNamespace);
+  return listScenarioIn(program, program.getGlobalNamespaceType());
 }
 
 export function getScenarioEndpoints(program: Program, target: Namespace | Interface | Operation): ScenarioEndpoint[] {
@@ -133,8 +129,27 @@ function getRouteSegments(program: Program, target: Operation | Interface | Name
 }
 
 function getOperationRoute(program: Program, target: Operation): string {
+  const template = getRouteSegmentFromServer(program);
   const segments = getRouteSegments(program, target);
-  return "/" + segments.map((x) => (x.startsWith("/") ? x.substring(1) : x)).join("/");
+  return (
+    (template ? (template.endsWith("/") || segments.length === 0 ? template : template + "/") : "/") +
+    segments.map((x) => (x.startsWith("/") ? x.substring(1) : x)).join("/")
+  );
+}
+
+function getRouteSegmentFromServer(program: Program): string | undefined {
+  const serviceNs = listServices(program)[0]?.type;
+  const server = getServers(program, serviceNs);
+  if (server && server.length === 1) {
+    if (server[0].url.indexOf("localhost:3000") > -1) {
+      return server[0].url.split("localhost:3000")[1];
+    } else if (server[0].url.indexOf("{endpoint}") > -1) {
+      return server[0].url.split("{endpoint}")[1];
+    } else {
+      return server[0].url;
+    }
+  }
+  return undefined;
 }
 
 export function listScenarioIn(program: Program, target: Namespace | Interface | Operation): Scenario[] {